+++ search-app/src/cmr/search/api/routes.clj
@@ -24,13 +24,13 @@
             [cmr.common.mime-types :as mt]
             [cmr.common.xml :as cx]
             [cmr.search.services.query-service :as query-svc]
-            [cmr.search.services.humanizers-service :as humanizers-service]
             [cmr.common.api.context :as context]
             [cmr.search.data.metadata-retrieval.metadata-cache :as metadata-cache]
             [cmr.search.services.parameters.legacy-parameters :as lp]
             [cmr.search.services.messages.common-messages :as msg]
             [cmr.search.services.health-service :as hs]
             [cmr.search.api.tags-api :as tags-api]
+            [cmr.search.api.humanizer :as humanizers-api]
             [cmr.umm-spec.versioning :as umm-version]
             [cmr.acl.core :as acl]
             [cmr.search.api.keyword :as keyword-api]
@@ -311,13 +311,6 @@
      :headers {cr/CONTENT_TYPE_HEADER (mt/with-utf-8 mt/json)}
      :body results}))
 
-(defn- humanizers-report
-  "Handles a request to get a humanizers report"
-  [context]
-  {:status 200
-   :headers {cr/CONTENT_TYPE_HEADER mt/csv}
-   :body (humanizers-service/humanizers-report-csv context)})
-
 (defn- build-routes [system]
   (let [relative-root-url (get-in system [:public-conf :relative-root-url])]
     (routes
@@ -325,6 +318,9 @@
 
         ;; Add routes for tagging
         tags-api/tag-api-routes
+
+        ;; Add routes for humanizers
+        humanizers-api/humanizers-routes
 
         ;; Add routes for API documentation
         (api-docs/docs-routes
@@ -340,10 +336,6 @@
 
         ;; Routes for collection html resources
         (collection-renderer-routes/resource-routes system)
-
-        (context "/humanizers" []
-          (GET "/report" {context :request-context}
-            (humanizers-report context)))
 
         ;; Retrieve by cmr concept id or concept id and revision id
         ;; Matches URL paths of the form /concepts/:concept-id[/:revision-id][.:format],