+++ umm-spec-lib/src/cmr/umm_spec/record_generator.clj
@@ -3,9 +3,9 @@
   (:require [cmr.umm-spec.json-schema :as js]
             [clojure.java.io :as io]
             [clojure.string :as str]
-            [cmr.umm-spec.models.collection]
-            [cmr.umm-spec.models.common]
-            [cmr.umm-spec.models.service]))
+            [cmr.umm-spec.models.umm-collection-models]
+            [cmr.umm-spec.models.umm-common-models]
+            [cmr.umm-spec.models.umm-service-models]))
 
 ;; Improvements
 ;; - generate records with fields in the same order as they are defined in the file.
@@ -13,9 +13,9 @@
 
 (def schema-name->namespace
   "A map of schema names to the namespace they should be placed in"
-  {"umm-cmn-json-schema.json" 'cmr.umm-spec.models.common
-   "umm-c-json-schema.json" 'cmr.umm-spec.models.collection
-   "umm-s-json-schema.json" 'cmr.umm-spec.models.service})
+  {"umm-cmn-json-schema.json" 'cmr.umm-spec.models.umm-common-models
+   "umm-c-json-schema.json" 'cmr.umm-spec.models.umm-collection-models
+   "umm-s-json-schema.json" 'cmr.umm-spec.models.umm-service-models})
 
 (defn schema-type-constructor
   "Returns the map->RecordName function that can be used to construct a type defined in the JSON
@@ -167,15 +167,15 @@
 (defn generate-umm-records
   "Generates all the UMM records"
   []
-  (generate-clojure-records-file {:the-ns 'cmr.umm-spec.models.common
+  (generate-clojure-records-file {:the-ns 'cmr.umm-spec.models.umm-common-models
                                   :description "Defines UMM Common clojure records."
                                   :schema-resource js/umm-cmn-schema-file})
 
-  (generate-clojure-records-file {:the-ns 'cmr.umm-spec.models.collection
+  (generate-clojure-records-file {:the-ns 'cmr.umm-spec.models.umm-collection-models
                                   :description "Defines UMM-C clojure records."
                                   :schema-resource (js/concept-schema-resource :collection)})
 
-  (generate-clojure-records-file {:the-ns 'cmr.umm-spec.models.service
+  (generate-clojure-records-file {:the-ns 'cmr.umm-spec.models.umm-service-models
                                   :description "Defines UMM-S clojure records."
                                   :schema-resource (js/concept-schema-resource :service)}))