+++ umm-spec-lib/test/cmr/umm_spec/test/validation/spatial.clj
@@ -2,7 +2,7 @@
   "This has tests for UMM validations."
   (:require [clojure.test :refer :all]
             [cmr.umm-spec.validation.umm-spec-validation-core :as v]
-            [cmr.umm-spec.models.collection :as coll]
+            [cmr.umm-spec.models.umm-collection-models :as coll]
             [cmr.umm-spec.test.validation.umm-spec-validation-test-helpers :as helpers]
             [cmr.spatial.mbr :as m]
             [cmr.spatial.point :as p]